refactor actions into submodules
